see if the GH build server passes these tests
